dags/requirements.txt
apache-airflow-providers-postgres
psycopg2-binary
pydantic-settings
requests
beautifulsoup4